test(talk): split digest lifecycle coverage
